Update AZ-300T01_Lab_Mod01_Exploring Monitoring Capabilities in Azure.md
Basically wanting to add in information regarding the fact that the action created in an earlier step was not actually mentioned in terms of being utilized when creating the alert.

I was going to mention that the Azure CLI resource group removal didn't work but after checking again it does seem to be correct syntax wise. However I had more joy when only specifying 'az30', not sure why or if it was just a one off when trying what was mentioned.

Also noticed that Line 199 and 209 do not get rid off all resource groups that would get created if this lab was followed to the letter.