Restructure due to Packer now in Cloud Shell
Cloud Shell now has Packer pre-installed as per https://azure.microsoft.com/en-us/updates/new-tools-in-azure-cloud-shell/ so there's no need to download and unzip it.  This means Exercise 1 can be removed, if the two "prerequisite" steps are moved to the next exercise.  Those two steps don't necessarily need to be split as the very next steps are similar command patterns.