Update README.md
for interaction between the course authors and MCTs. We hope this will help  keep the = for interaction between the course authors and MCTs. We hope this will help keep the
content current as the Azure platform changes =  keep the content current as of the Azure platform changes.
Architect Technologies , and AZ-301, Microsoft Azure Architect Design, = Architect Technologies, and AZ-301, Microsoft Azure Architect Design